refactor(repair): move masterSelector to a separate file
This change increases readability.
