+++ pkg/service/repair/generator.go
@@ -4,8 +4,6 @@ package repair
 
 import (
 	"context"
-	"math"
-	"sort"
 	"strings"
 	"sync/atomic"
 
@@ -14,46 +12,7 @@ import (
 	"github.com/scylladb/go-set/strset"
 	"github.com/scylladb/scylla-manager/v3/pkg/dht"
 	"github.com/scylladb/scylla-manager/v3/pkg/scyllaclient"
-	"github.com/scylladb/scylla-manager/v3/pkg/util/slice"
 )
-
-// masterSelector describes each host priority for being repair master.
-// Repair master is first chosen by smallest shard count,
-// then by smallest dc RTT from SM.
-type masterSelector map[string]int
-
-func newMasterSelector(shards map[string]uint, hostDC map[string]string, closestDC []string) masterSelector {
-	hosts := make([]string, 0, len(shards))
-	for h := range shards {
-		hosts = append(hosts, h)
-	}
-
-	sort.Slice(hosts, func(i, j int) bool {
-		if shards[hosts[i]] != shards[hosts[j]] {
-			return shards[hosts[i]] < shards[hosts[j]]
-		}
-		return slice.Index(closestDC, hostDC[hosts[i]]) < slice.Index(closestDC, hostDC[hosts[j]])
-	})
-
-	ms := make(masterSelector)
-	for i, h := range hosts {
-		ms[h] = i
-	}
-	return ms
-}
-
-// Select returns repair master from replica set.
-func (ms masterSelector) Select(replicas []string) string {
-	var master string
-	p := math.MaxInt64
-	for _, r := range replicas {
-		if ms[r] < p {
-			p = ms[r]
-			master = r
-		}
-	}
-	return master
-}
 
 type submitter[T, R any] interface {
 	Submit(task T)
